requirements.txt
Flask==1.1.1
Flask-Caching
gunicorn
requests
spotipy
python-dotenv
markdown
PyYAML
python-frontmatter
boto3==1.13.24